refactor: apply-feedback
1. Long타입 사용
2. Line-Section 양방향 연관관계 사용
3. 파라미터 검증 도메인 객체로 이동
이전 단계 피드백을 적용했습니다.
