+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,6 +1,7 @@
-package nextstep.subway.domain;
+package nextstep.subway.line.domain;
 
 import lombok.Getter;
+import org.springframework.util.StringUtils;
 
 import javax.persistence.*;
 import java.util.Collections;
@@ -28,22 +29,32 @@ public class Line {
     }
 
     public void addSection(Long upStationId, Long downStationId, int distance) {
-        sections.add(new Section(upStationId, downStationId, distance));
+        sections.add(new Section(this, upStationId, downStationId, distance));
     }
 
-    public void removeSection(long stationId) {
+    public void removeSection(Long stationId) {
         sections.removeSection(stationId);
     }
 
     public void updateName(String name) {
+        if (!StringUtils.hasText(name)) {
+            throw new IllegalArgumentException("노선의 이름은 필수값입니다.");
+        }
         this.name = name;
     }
 
     public void updateColor(String color) {
+        if (!StringUtils.hasText(color)) {
+            throw new IllegalArgumentException("노선의 색상은 필수값입니다.");
+        }
         this.color = color;
     }
 
     public List<Section> getSections() {
         return Collections.unmodifiableList(sections.getSections());
+    }
+
+    public List<Long> stationIds() {
+        return sections.stationIds();
     }
 }
