+++ src/main/java/nextstep/subway/line/domain/Section.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.line.domain;
 
 import lombok.Getter;
 
@@ -11,6 +11,10 @@ public class Section {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "line_id")
+    private Line line;
+
     private Long upStationId;
 
     private Long downStationId;
@@ -20,7 +24,8 @@ public class Section {
     protected Section() {
     }
 
-    public Section(Long upStationId, Long downStationId, int distance) {
+    public Section(Line line, Long upStationId, Long downStationId, int distance) {
+        this.line = line;
         this.upStationId = upStationId;
         this.downStationId = downStationId;
         this.distance = distance;
