+++ src/main/java/nextstep/subway/line/domain/Sections.java
@@ -1,28 +1,27 @@
-package nextstep.subway.domain;
+package nextstep.subway.line.domain;
 
 import lombok.Getter;
-import nextstep.subway.domain.exception.CannotDeleteSectionException;
+import nextstep.subway.line.domain.exception.CannotDeleteSectionException;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Embeddable;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToMany;
+import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static javax.persistence.CascadeType.*;
 
 @Getter
 @Embeddable
 public class Sections {
 
-    @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
-    @JoinColumn(name = "line_id")
+    @OneToMany(mappedBy = "line", cascade = ALL, orphanRemoval = true)
     private List<Section> sections = new ArrayList<>();
 
     public void add(Section section) {
         sections.add(section);
     }
 
-    public void removeSection(long stationId) {
+    public void removeSection(Long stationId) {
         if (sections.isEmpty()) {
             throw new CannotDeleteSectionException("현재 존재하는 구간이 없습니다.");
         }
@@ -36,5 +35,18 @@ public class Sections {
 
     private Section lastSection() {
         return sections.get(sections.size() - 1);
+    }
+
+    public List<Long> stationIds() {
+        List<Long> stationIds = sections.stream()
+                .map(Section::getDownStationId)
+                .collect(Collectors.toList());
+
+        stationIds.add(0, firstSection().getUpStationId());
+        return stationIds;
+    }
+
+    private Section firstSection() {
+        return sections.get(0);
     }
 }
